types: extract page response types to page.ts
